kernel/pyproject.toml
[project]
name = "cyntra"
version = "0.1.0"
description = "Local-first autonomous kernel using Beads + Workcells"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Connor", email = "connor@example.com" }]
keywords = ["ai", "agents", "orchestration", "development", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "structlog>=24.0.0",
    "aiofiles>=24.0.0",
    "gitpython>=3.1.0",
    "jsonschema>=4.20.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
]

evolve = [
    "dspy-ai>=2.4.0",
]

fab = [
    # Core ML dependencies for critics
    "torch>=2.0.0",
    "open-clip-torch>=2.24.0",
    "Pillow>=10.0.0",
    # Mesh analysis
    "trimesh>=4.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
]

fab-cpu = [
    # CPU-only torch for deterministic inference
    "torch>=2.0.0",
    "open-clip-torch>=2.24.0",
    "Pillow>=10.0.0",
    "trimesh>=4.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
]

memory = [
    # PostgreSQL with pgvector for embeddings
    "asyncpg>=0.29.0",
    "pgvector>=0.2.0",
    # Embedding generation
    "sentence-transformers>=2.2.0",
    # LLM client for extraction/consolidation
    "anthropic>=0.25.0",
    # Vector operations
    "numpy>=1.24.0",
    # Token counting (optional, for accurate context limits)
    "tiktoken>=0.5.0",
]

indexing = [
    # CocoIndex indexing + query handlers
    "cocoindex[embeddings]>=0.3.22,<0.4",
    # Some embedding models require this at runtime (e.g. nomic-embed).
    "einops>=0.8.0",
    # Postgres query-time vector support (psycopg + pgvector)
    "pgvector>=0.2.0",
    "psycopg[binary]>=3.2.0",
    "psycopg_pool>=3.2.0",
]

[project.scripts]
cyntra = "cyntra.cli:main"
workcell = "cyntra.workcell.cli:main"
fab-gate = "cyntra.fab.gate:main"
fab-render = "cyntra.fab.render:main"
fab-godot = "cyntra.fab.godot:main"
fab-critics = "cyntra.fab.critics.cli:main"
fab-regression = "cyntra.fab.regression:main"
fab-scaffold = "cyntra.fab.scaffolds.cli:main"
fab-world = "cyntra.fab.world:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cyntra"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "cocoindex",
    "cocoindex.*",
    "numpy",
    "numpy.*",
    "pgvector",
    "pgvector.*",
    "psycopg_pool",
    "psycopg_pool.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/cyntra"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
